+++ BeatBoxer-Tui/src/app/render/mod.rs
@@ -4,3 +4,4 @@ mod render_manage_section;
 mod render_render_section;
 mod render_buttons_section;
 mod render_utils_section;
+mod render_wave_form;